# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mint Bucks promotional credit system for Mint Printworks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: hasCredit
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /customers/{id}/credits:
    get:
      operationId: getCustomerCredits
      tags: [customers]
      summary: Get all credits for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer credits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credit"

  # ── Credits ────────────────────────────────────────────────────────────────
...
  /redemptions:
    get:
      operationId: listRedemptions
      tags: [redemptions]
      summary: List all redemptions
      parameters:
        - name: customerId
          in: query
          schema: { type: string }
        - name: creditId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of redemptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Redemption"
  /redemptions/{id}:
    get:
      operationId: getRedemption
      tags: [redemptions]
      summary: Get redemption by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Redemption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redemption"
        "404":
          description: Not found

  # ── Reports ────────────────────────────────────────────────────────────────
...
  /printavo/test:
    post:
      operationId: testPrintavoConnection
      tags: [printavo]
      summary: Test Printavo API connection
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrintavoTestInput"
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintavoConnectionResult"
  /printavo/sync-customers:
    post:
      operationId: syncPrintavoCustomers
      tags: [printavo]
      summary: Pull customers from Printavo and upsert into local table
      responses:
        "200":
          description: Sync summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintavoSyncResult"
        "400":
          description: Printavo not configured
...
  /printavo/order/{orderNumber}:
    get:
      operationId: lookupPrintavoOrder
      tags: [printavo]
      summary: Fetch a Printavo order by order number for reference
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintavoOrderSummary"
        "404":
          description: Order not found
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, email, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        outstandingBalance:
          type: number
        totalIssued:
          type: number
        totalRedeemed:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
    CustomerInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        companyName:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
    Credit:
      type: object
      required: [id, customerId, customerName, code, amount, amountRemaining, status, issuedAt, createdAt, updatedAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerCompany:
          type: ["string", "null"]
        code:
          type: string
        amount:
          type: number
        amountRemaining:
          type: number
        status:
          type: string
          enum: [active, partially_redeemed, redeemed, expired, cancelled]
        note:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        issuedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Redemption:
      type: object
      required: [id, creditId, customerId, customerName, amountApplied, redeemedAt, createdAt]
      properties:
        id:
          type: integer
        creditId:
          type: integer
        creditCode:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        customerCompany:
          type: ["string", "null"]
        amountApplied:
          type: number
        invoiceRef:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        redeemedAt:
          type: string
        createdAt:
          type: string
...
    PrintavoTestInput:
      type: object
      properties:
        apiKey:
          type: string
        email:
          type: string
...
    PrintavoConnectionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    PrintavoSyncResult:
      type: object
      required: [created, matched, skipped, total]
      properties:
        created:
          type: integer
        matched:
          type: integer
        skipped:
          type: integer
        total:
          type: integer
    PrintavoOrderSummary:
      type: object
      required: [id, visualId, createdAt]
      properties:
        id:
          type: string
        visualId:
          type: string
        orderId:
          type: ["string", "null"]
        createdAt:
          type: string
        total:
          type: ["number", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
...